src/azure-ai-foundry-memory/requirements.txt
agent-framework==1.0.0
azure-ai-projects
azure-identity
python-dotenv
